changed int to uint in test (avoid zero sum)

+++ CROSS_submission_package_v1.2/Additional_Implementations/test_speed/extra/OK_speed.c
@@ -57,7 +57,7 @@ int main() {
     clock_t t_ope = 0;
     clock_t t_tmp = 0;
 
-    int errors = 0;
+    uint16_t errors = 0;
 
     for(int i=0; i<NUM_TESTS; i++) {
